# openapi.yaml
openapi: 3.0.0
info:
  version: '1.1'
  title: Ruokavälitys API
  description: API for both, managing the RV system and it's inventory, and browsing and purchasing products.
...
servers:
  - url: https://rv-backend-dev.herokuapp.com/
    description: Staging environment
  - url: https://rv-backend.herokuapp.com/
    description: Production environment
...
components:
  securitySchemes:
    normalAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the \"end user facing\" portition of the API.
        Can be obtained with any valid credentials from route `/api/v1/authenticate`.
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the administrations portition of the API.
        Can be obtained with administrator credentials from route `/api/v1/admin/authenticate`.
  schemas:
    Role:
      type: string
      enum:
        - ADMIN
        - USER1
        - USER2
        - INACTIVE
      example: USER
    Product:
      type: object
      required: 
        - barcode
        - name
        - category
        - sellPrice
        - stock
      properties: 
        barcode: 
          type: string
          example: "6430039222212"
          description: Barcode on the product packaging. Used to uniquely identify the product.
        name: 
          type: string
          example: Kombucha bakteerin makuinen bakteerijuoma
          description: Human readable and user facing name of the product.
        category: 
          '$ref': '#/components/schemas/Category'
        sellPrice: 
          type: integer
          example: 310
          description: Price (in cents) at which the product is on sale.  
        stock: 
          type: integer
          example: 10
          description: Number of items in stock.
...
    PurchaseEventBase:
      type: object
      required: 
        - purchaseId
        - time
        - price
      properties: 
        purchaseId: 
          type: integer
          example: 3053
          description: Opaque unique identifier of the purchase event.
        time: 
          type: string
          format: date-time
          example: '2018-12-24T00:00:00.000Z'
          description: Timestamp of the event in ISO-8601 format.
        price: 
          type: integer
          example: 19
          description: Price (in cents) of the product bought.
    PurchaseEventProduct:
      type: object
      required:
        - product
      properties:
        product:
          '$ref': '#/components/schemas/Product'
...
    PurchaseEventBalance:
      type: object
      required:
        - balanceAfter
      properties:
        balanceAfter:
          type: integer
          example: 1981
          description: Balance of the user's account (in cents) after the purchase transaction.
...
    DepositEventBase:
      type: object
      required: 
        - depositId
        - time
        - amount
        - balanceAfter
      properties: 
        depositId: 
          type: integer
          description: Opaque unique identifier of the deposit event.
        time: 
          type: string
          format: date-time
          example: '2018-12-24T00:00:00.000Z'
          description: Timestamp of the event in ISO-8601 format.
        amount: 
          type: integer
          example: 2000
          description: Amount deposited to the user's account in cents.
        balanceAfter:
          type: integer
          example: 2500
          description: Balance of the user's account after the deposit in cents.
...
    User: 
      type: object
      required: 
        - userId
        - username
        - email
        - moneyBalance
        - role
        - privacyLevel
      properties: 
        userId: 
          type: integer
          example: 415
          description: Opaque unique identifier of the user.
        username: 
          type: string
          example: fuxi2001
          description: Username which is used in authentication.
        fullName:
          type: string
          nullable: true
          example: Fuxi C. S. McStudent
          description: Real full name of the user.
        email: 
          type: string
          example: fuxi.student@helsinki.fi
          description: E-mail address at which the user can be reached.
        moneyBalance: 
          type: integer
          example: 2000
          description: Balance of the user's account in cents.
        privacyLevel:
          type: integer
          example: 0
          description: Privacy level of the user, 0 = no restrictions, 1 = hide username, 2 = hide even anonymous data
          enum: [0, 1, 2]
        role:
          '$ref': '#/components/schemas/Role'
    Category:
      type: object
      required: 
        - categoryId
        - description
      properties: 
        categoryId: 
          type: integer
          example: 10
          description: Opaque unique identifier of the category.
        description: 
          type: string
          example: Drinks, soda with bottle return
          description: Human readable and user facing description of the product category.
...
  responses:
...
    NotAuthorizedErrorResponse:
      description: Your user lacks the required privilege level to perform this action.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_authorized
                  - invalid_token
              message:
                type: string
                example: Not authorized
    NotFoundErrorResponse:
      description: The resource your are requesting does not exist.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_found
              message:
                type: string
                example: No product with barcode \'1312767817\' found
    BadRequestErrorResponse:
      description: The request you sent contains invalid values or is missing fields.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - bad_request
                  - invalid_reference
              message:
                type: string
                example: Missing or invalid fields in request
              errors:
                type: array
                items:
                  type: string
                  example: 'boxBarcode should be a numeric 1-14 digit barcode' 
    IdentifierTakenErrorResponse:
      description: Creating a new resource failed, because a resource with the given identifier already exists.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - identifier_taken
              message:
                type: string
                example: Barcode already in use.
paths: 
...
  /api/v1/user:
    get: 
      summary: Get details of the authenticated user
      operationId: getSessionUser
      tags:
        - normal
      security:
        - normalAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - user
                properties: 
                  user:
                    '$ref': '#/components/schemas/User'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
    patch: 
      summary: Update details of the authenticated user
      operationId: updateSessionUser
      tags:
        - normal
      security:
        - normalAccessToken: []
      responses: 
        200: 
          description: User information updated successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - user
                properties: 
                  user: 
                    '$ref': '#/components/schemas/User'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        409:
          '$ref': '#/components/responses/IdentifierTakenErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              minProperties: 1
              properties: 
                username: 
                  type: string
                  example: 'jäärä2001'
                  minLength: 1
                fullName: 
                  type: string
                  example: 'Jäärä F. S. McConsultant'
                  minLength: 1
                email: 
                  type: string
                  example: 'jaara.fs.constultant@coolcorp.com'
                  minLength: 1
...
  /api/v1/user/changePassword:
    post:
      summary: Change the account's password
      operationId: changePassword
      tags:
        - normal
      security:
        - normalAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  example: hunter2
                  minLength: 1
      responses:
        204:
          description: Password changed successfully
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
...
  /api/v1/admin/users: 
    get: 
      summary: Get a list of all users
      operationId: getUsers
      tags:
        - admin
      security:
        - adminAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - users
                properties: 
                  users: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/User'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
  /api/v1/admin/users/{userId}: 
    get: 
      summary: Get the details of a user
      operationId: getUser
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: userId
          description: Unique identifier of the user.
          required: true
          in: path
          schema: 
            type: integer
            example: 324
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - user
                properties: 
                  user: 
                    '$ref': '#/components/schemas/User'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
  /api/v1/admin/users/{userId}/purchaseHistory: 
    get: 
      summary: Get user's purchase history
      operationId: getUserPurchaseHistory
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: userId
          description: Unique identifier of the user.
          required: true
          in: path
          schema: 
            type: integer
            example: 324
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - purchases
                properties: 
                  purchases: 
                    type: array
                    items: 
                      allOf:
                          - '$ref': '#/components/schemas/PurchaseEventBase'
                          - '$ref': '#/components/schemas/PurchaseEventProduct'
                          - '$ref': '#/components/schemas/PurchaseEventBalance'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/admin/users/{userId}/depositHistory: 
    get: 
      summary: Get deposit history of a user
      operationId: getUserDepositHistory
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: userId
          description: Unique identifier of the user.
          required: true
          in: path
          schema: 
            type: integer
            example: 324
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deposits
                properties: 
                  deposits: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/DepositEventBase'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...